implement 99bill payment service
