+++ payment-runtime/src/main/java/org/klose/payment/integration/bill99/Bill99PaymentService.java
@@ -8,6 +8,7 @@ import org.klose.payment.common.utils.Assert;
 import org.klose.payment.common.utils.LogUtils;
 import org.klose.payment.constant.FrontPageForwardType;
 import org.klose.payment.constant.PaymentConstant;
+import org.klose.payment.integration.bill99.constant.Bill99Constant;
 import org.klose.payment.service.AccountService;
 import org.klose.payment.service.PaymentService;
 import org.klose.payment.service.TransactionDataService;
@@ -36,101 +37,7 @@ public class Bill99PaymentService implements PaymentService {
     @Autowired
     private AccountService accountService;
 
-
-    private Logger logger = LoggerFactory.getLogger(this.getClass());
-
-    final static String NOTIFY_RESPONSE_MSG = "<result>%d</result><redirecturl>%s?channelCode=%s&quotationNumber=%s</redirecturl>";
-
-    final static String PARAM_KEY_RESULT = "payResult";
-
-    final static String PARAM_KEY_DEAL_ID = "dealId";
-
-    final static String PARAM_KEY_ORDER_ID = "orderId";
-
-    final static String PARAM_KEY_CHANNEL_CODE = "ext1";
-
-    final static String PARAM_KEY_PRODUCT_ID = "ext2";
-
-    final static String PAYMENT_SUCCESS_CODE = "10";
-
-    public JSONObject buildPaymentRequestData(String proxyId,  HttpServletRequest request) throws Exception {
-
-        JSONObject requestData = new JSONObject();
-
-//        String orderId = policy.getQuotationNumber();
-//        String orderAmount = String.valueOf(policy.getFee("AGP").getAmount().multiply(new BigDecimal("100")).intValue());
-//        String productId = policy.getProductId();
-//
-//        String redirectURL = PaymentUtility.getKuaiqianPaymentRedirectURL(request.getRequestURL().toString(), productId);
-//        String callbackURL = PaymentUtility.getKuaiqianPaymentNotificationURL(request.getRequestURL().toString());
-//
-//        String productCode = policy.getProductName();
-//        String channelCode = policy.getChannelCode();
-//        Map<String, String> params = kqHelper.buildPaymentRequestParams(
-//                orderId, orderAmount.toString(),
-//                redirectURL,
-//                callbackURL,
-//                productCode, orderAmount.toString(), productCode, productId, channelCode);
-//        String signature = kqHelper.sign(params);
-//        requestData.putAll(params);
-//        requestData.put("signMsg", signature);
-//        requestData.put("paymentEndpoint", String.format("%s/recvMerchantInfoAction.htm", kqHelper.getGatewayURL()));
-//
-//        PaymentData paymentData = new PaymentData();
-//        paymentData.setOrderNumber(orderId);
-//        paymentData.setAmount(policy.getFee("AGP").getAmount());
-//        paymentData.setSignString(signature);
-//        paymentData.setStatus(PaymentData.STATUS_PROCEED);
-//        paymentDataDao.save(paymentData);
-
-        return requestData;
-    }
-
-
-
-    public Response handleCallback(String proxyId, Map params, String encoding) {
-
-
-//        String paymentResult = ((String) params.get(PARAM_KEY_RESULT)).equals(PAYMENT_SUCCESS_CODE) ? "SUCCESS" : "FAIL";
-//        String quoteNumber = (String) params.get(PARAM_KEY_ORDER_ID);
-//        String payId = (String) params.get(PARAM_KEY_DEAL_ID);
-//        String channelCode = (String) params.get(PARAM_KEY_CHANNEL_CODE);
-//        String productId = (String) params.get(PARAM_KEY_PRODUCT_ID);
-//
-//        boolean verified = kqHelper.verify(params);
-//
-//        if (verified && quoteNumber != null) {
-//
-//            IndividualAgreement policy = insuranceAgreementService.findByQuotationNumber(quoteNumber);
-//
-//            try {
-//                if ("SUCCESS".equals(paymentResult) && policy != null) {
-//                    /*List<PrePay> prePays = policy.getChildrenByType(PrePay.class);
-//					if(prePays.size() > 0){
-//						PrePay prePay = prePays.get(0);
-//						prePay.setStatus(PrePayStatus.PAYMENT_SUCCESS);
-//						prePay.setReceiptDate(new Date());
-//					}*/
-//                    if (callBackService.processPaymentNotification(paymentResult, quoteNumber, payId)) {
-//                        callBackService.processPolicy(quoteNumber, paymentResult, payId);
-//                    }
-//                }
-//            } catch (Exception e) {
-//                policy.setStatusCode(AgreementStatusCodeList.WAITING_MANUAL_HANDLING);
-//                insuranceAgreementService.savePolicy(policy);
-//            }
-//        }
-//
-//        String return_msg = String.format(
-//                NOTIFY_RESPONSE_MSG,
-//                verified ? 1 : 0,
-//                PaymentUtility.getKuaiqianPaymentRedirectURL((String) params.get("requestURL"), productId),
-//                channelCode,
-//                quoteNumber);
-        String return_msg = "";
-        logger.info("KQ Return MSG: " + return_msg);
-        return Response.status(Response.Status.OK).entity(return_msg).type(MediaType.APPLICATION_XML_TYPE).build();
-    }
+    private Logger logger = LoggerFactory.getLogger(Bill99PaymentService.class);
 
 
     @Override
@@ -150,8 +57,7 @@ public class Bill99PaymentService implements PaymentService {
 
         String gateway = (String) bill.getConfigData().get("gateway");
         Assert.isNotNull(gateway);
-        String gatewayUrl = gateway;
-        paymentForm.setGatewayURL(gatewayUrl);
+        paymentForm.setGatewayURL(gateway);
 
         paymentForm.setParams(this.generatePaymentForm(transId, bill));
 
@@ -175,20 +81,21 @@ public class Bill99PaymentService implements PaymentService {
     private Map<String, String> generatePaymentForm(Long transId, BillingData bill) {
         Map<String, String> params = new LinkedHashMap<>();
 
-        params.put("inputCharset", Bill99Helper.INPUT_CHARSET);
+        params.put("inputCharset", Bill99Constant.INPUT_CHARSET);
+
+        params.put("version", Bill99Constant.VERSION);
+        params.put("language", Bill99Constant.LANGUAGE);
+        params.put("signType", Bill99Constant.SIGN_TYPE);
+
         // 服务器接收分账结果的后台地址
         String hostEndpoint = bill.getContextPath();
         Assert.isNotNull(hostEndpoint);
         final String PATH_NOTIFY_API = "/api/99bill/notifications";
         params.put("bgUrl", hostEndpoint.concat(PATH_NOTIFY_API));
-
         // 服务器回调前台地址
         String pageUrl = hostEndpoint.concat(PaymentConstant.GENERAL_RETURN_PROXY_PATH)
                 .concat(transId.toString());
 
-        params.put("version", Bill99Helper.VERSION);
-        params.put("language", Bill99Helper.LANGUAGE);
-        params.put("signType", Bill99Helper.SIGN_TYPE);
         params.put("merchantAcctId", (String) bill.getConfigData().get("merchantAcctId"));
 
         params.put("orderId", bill.getBizNo()); // quotation number
@@ -197,15 +104,15 @@ public class Bill99PaymentService implements PaymentService {
         params.put("orderTime", DateUtil.format(new Date(),
                 DateUtil.DATE_TIME_FORMAT_COMPACT_S));
 
-        //@TODO
-//        params.put("productName", productName);
-//        params.put("productNum", productNum);
-//        params.put("productDesc", productDesc);
-//
-//        params.put("ext1", channelCode);
-//        params.put("ext2", productId);
+        params.put("productName", bill.getSubject());
+        params.put("productNum", String.valueOf(bill.getQuantity()));
+        params.put("productDesc", bill.getDescription());
+        params.put("ext1", (String) bill.getExtData().get("channelCode"));
+        params.put("ext2", (String) bill.getExtData().get("productId"));
+        params.put("payType", Bill99Constant.PAY_TYPE);
 
-        params.put("payType", Bill99Helper.PAY_TYPE);
+//        requestData.put("signMsg", signature);
+
         return params;
     }
 }
