+++ payment-runtime/src/main/java/org/klose/payment/integration/bill99/rest/Bill99NotificationResource.java
@@ -1,8 +1,8 @@
 package org.klose.payment.integration.bill99.rest;
 
 import org.klose.payment.integration.bill99.Bill99Helper;
-import org.klose.payment.integration.bill99.Bill99PaymentService;
-import org.klose.payment.service.PaymentService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -18,54 +19,18 @@ import java.util.Map;
 @Path("bill99")
 @Component
 public class Bill99NotificationResource {
-
-	
 	@Autowired
 	private Bill99Helper helper;
-	
-	@Autowired
-	private Bill99PaymentService paymentService;
-	
-//	private final static Logger log = LoggerFactory.getLogger(this.getClass());
 
-	
-//	@GET
-//	@Path("/kuaiqian/endpoints")
-//	@Produces(MediaType.APPLICATION_JSON)
-//	public Response createKuaiQianPaymentParameters(
-//			@Context HttpServletRequest request,
-//			@QueryParam("orderId") String orderId,
-//			@QueryParam("orderAmount") String orderAmount,
-//			@QueryParam("policyNo") String policyNo,
-//			@QueryParam("callbackURL") String callbackURL) throws UnsupportedEncodingException {
-//
-//		log.debug(String.format("KuaiQianPayment order request: %s callbackURL: %s", orderId, callbackURL));
-//
-//		IndividualAgreement policy = agreementService.retrieve(policyNo, null);
-//
-//		String productId = policy.getProductId();
-//		String redirectURL = PaymentUtility.getKuaiqianPaymentRedirectURL(request.getRequestURL().toString(), productId);
-//
-//		Map<String, String> params = kqHelper.buildPaymentRequestParams(
-//				orderId, orderAmount, redirectURL, callbackURL,
-//				policy.getProductName(), orderAmount, policy.getProductName(),
-//				policy.getProductId(), policy.getChannelCode());
-//		String signature = kqHelper.sign(params);
-//
-//		params.put("signMsg", signature);
-//		params.put("paymentEndpoint", String.format("%s/recvMerchantInfoAction.htm", kqHelper.getGatewayURL()));
-//
-//		return Response.ok(BeanHelper.toJSON(params), MediaType.APPLICATION_JSON).build();
-//	}
+	private final static Logger logger = LoggerFactory.getLogger(Bill99NotificationResource.class);
+
 	
 	@SuppressWarnings("unchecked")
 	@GET
 	@Path("/callback")
-	public Response processKuaiQianNotifcation(@Context HttpServletRequest request) {
-	
+	public Response processNotification(@Context HttpServletRequest request) {
 		@SuppressWarnings("rawtypes")
 		Enumeration keys = request.getParameterNames();
-		
 		Map<String, Object> params = new HashMap<>();
 		while(keys.hasMoreElements()) {
 			String key = (String)keys.nextElement();
@@ -74,7 +39,50 @@ public class Bill99NotificationResource {
 		
 		params.put("requestURL", request.getRequestURL().toString());
 			
-		return paymentService.handleCallback("kuaiqian", params, request.getCharacterEncoding());
+		return handleCallback("kuaiqian", params, request.getCharacterEncoding());
 	}
-	
+
+	public Response handleCallback(String proxyId, Map params, String encoding) {
+
+
+//        String paymentResult = ((String) params.get(PARAM_KEY_RESULT)).equals(PAYMENT_SUCCESS_CODE) ? "SUCCESS" : "FAIL";
+//        String quoteNumber = (String) params.get(PARAM_KEY_ORDER_ID);
+//        String payId = (String) params.get(PARAM_KEY_DEAL_ID);
+//        String channelCode = (String) params.get(PARAM_KEY_CHANNEL_CODE);
+//        String productId = (String) params.get(PARAM_KEY_PRODUCT_ID);
+//
+//        boolean verified = kqHelper.verify(params);
+//
+//        if (verified && quoteNumber != null) {
+//
+//            IndividualAgreement policy = insuranceAgreementService.findByQuotationNumber(quoteNumber);
+//
+//            try {
+//                if ("SUCCESS".equals(paymentResult) && policy != null) {
+//                    /*List<PrePay> prePays = policy.getChildrenByType(PrePay.class);
+//					if(prePays.size() > 0){
+//						PrePay prePay = prePays.get(0);
+//						prePay.setStatus(PrePayStatus.PAYMENT_SUCCESS);
+//						prePay.setReceiptDate(new Date());
+//					}*/
+//                    if (callBackService.processPaymentNotification(paymentResult, quoteNumber, payId)) {
+//                        callBackService.processPolicy(quoteNumber, paymentResult, payId);
+//                    }
+//                }
+//            } catch (Exception e) {
+//                policy.setStatusCode(AgreementStatusCodeList.WAITING_MANUAL_HANDLING);
+//                insuranceAgreementService.savePolicy(policy);
+//            }
+//        }
+//
+//        String return_msg = String.format(
+//                NOTIFY_RESPONSE_MSG,
+//                verified ? 1 : 0,
+//                PaymentUtility.getKuaiqianPaymentRedirectURL((String) params.get("requestURL"), productId),
+//                channelCode,
+//                quoteNumber);
+		String return_msg = "";
+		logger.info("KQ Return MSG: " + return_msg);
+		return Response.status(Response.Status.OK).entity(return_msg).type(MediaType.APPLICATION_XML_TYPE).build();
+	}
 }
