+++ payment-runtime/src/test/resources/99bill.cert.rsa.20340630.cer
@@ -1,0 +1,26 @@
+-----BEGIN CERTIFICATE-----
+MIIEVzCCAz+gAwIBAgIJAMyX6hgX0bZ3MA0GCSqGSIb3DQEBBQUAMHoxCzAJBgNV
+BAYTAkNOMQswCQYDVQQIEwJTSDELMAkGA1UEBxMCU0gxDzANBgNVBAoTBjk5Ymls
+bDEMMAoGA1UECxMDUE9BMQwwCgYDVQQDEwNJTkYxJDAiBgkqhkiG9w0BCQEWFXBv
+YV9lYV9pbmZAOTliaWxsLmNvbTAeFw0xNDA2MzAwMjMxMzBaFw0zNDA2MzAwMjMx
+MzBaMHoxCzAJBgNVBAYTAkNOMQswCQYDVQQIEwJTSDELMAkGA1UEBxMCU0gxDzAN
+BgNVBAoTBjk5YmlsbDEMMAoGA1UECxMDUE9BMQwwCgYDVQQDEwNJTkYxJDAiBgkq
+hkiG9w0BCQEWFXBvYV9lYV9pbmZAOTliaWxsLmNvbTCCASIwDQYJKoZIhvcNAQEB
+BQADggEPADCCAQoCggEBAKdZsacD41SXYXst4dfRR1mOIDKRvPb9G19SDT8TgTYL
+5CU0x+22Kxu0XhBWQ91iOtXaYF24HOoUnuZIC1DL1ul6WJoEnvGIdn3wXkh+jt6k
+whr/EAzBA2a8oBrDXRAXV5tb6AO90kZvz/H3hUW8zx8SEHm7pATuSvXgCSBWU6J2
+DGNfpcaxkj+G8PhrhfWgmy0J18+HbO9ugt9i4wV9S7VDKqgrcCb3K7VRZRiLpa5K
+vB+b+fHe4M15ODCSSD4wFPFxxWX1x4Ih8I25Bgrmx3tskWStYNbPDH4JW6ZpA19w
+fFtqA7eO12GuuQWdEGCtQSqxvgdKGoG1T02DdMWdlekCAwEAAaOB3zCB3DAdBgNV
+HQ4EFgQUr+b69rLVT8kUz5mOAtFXH2vqc1swgawGA1UdIwSBpDCBoYAUr+b69rLV
+T8kUz5mOAtFXH2vqc1uhfqR8MHoxCzAJBgNVBAYTAkNOMQswCQYDVQQIEwJTSDEL
+MAkGA1UEBxMCU0gxDzANBgNVBAoTBjk5YmlsbDEMMAoGA1UECxMDUE9BMQwwCgYD
+VQQDEwNJTkYxJDAiBgkqhkiG9w0BCQEWFXBvYV9lYV9pbmZAOTliaWxsLmNvbYIJ
+AMyX6hgX0bZ3MAwGA1UdEwQFMAMBAf8wDQYJKoZIhvcNAQEFBQADggEBAHSm2rzE
+oygqrOMKPypyMWd9LWXa8wvNkzBiQyuPbaIjx4BNZd8ImlaOCXgk5Z5JDja1jPyY
+Sac+G3tzk4e/9wUgRY8rxnKhr7NCPKdbWZGC5O27LM/mCM9wj+cWIvi1DKSD3oe/
+tkN26mbtcHpug8Ubzz9hABc7f07lZEe8tgV8bxvrLHzmFrEoBRerAKqMyGEb+a57
+RLa/Lv8/OhqRlYzERy6N32egpSCQVtWLMyQI861N/QqyKW3egwUsPLk5c6/qVsz2
+a4Ka5SUtsLYUS/tfo4zIitfXxDGqG0oYlFcthVu6iy6Pju5q5l/XDGnjTVHiiVmo
+1WeSU3Ps23gEhFM=
+-----END CERTIFICATE-----
